Project structure changed v2 1. extracted authorisation_view,authorisation_view_controller 2. extracted main_view, main_view_controller 3. extracted patient_screen, patients_screen_controller 4. cim_connection moved to cim_client\lib 5. CIManagerApp renamed to CIMApp 6. snake case implemented on file names 7. CIMUser and CIMPatient moved to cim_protocol 8. globals.dart refactored. Added enum CIMErrors for errors, keys from map with errors changed to CIMErrors
