--- cim_client/lib/CIMUser.dart
@@ -1,7 +1,0 @@
-class CIMUser{
-  String _login;
-  String get login => _login;
-  String _password;
-
-  CIMUser (this._login, this._password);
-}
