+++ cim_client/lib/views/connection_view.dart
@@ -1,7 +1,7 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'ConnectionViewController.dart';
+import 'connection_view_controller.dart';
 import 'package:get/get.dart' hide Trans;
 import 'package:easy_localization/easy_localization.dart';
