+++ cim_client/lib/views/main_view_controller.dart
@@ -1,7 +1,7 @@
-import 'package:cim_client/CIMDataProvider.dart';
-import 'package:cim_client/CIMUser.dart';
-import 'package:cim_client/views/cim_connection.dart';
-import 'package:cim_client/views/get_view_model.dart';
+import 'package:cim_client/cim_data_provider.dart';
+import 'package:cim_protocol/cim_protocol.dart';
+import 'package:cim_client/cim_connection.dart';
+import 'package:cim_client/views/patients_screen_controller.dart';
 import 'package:get/get.dart' hide Trans;
 import 'package:cim_client/globals.dart';
 import 'package:easy_localization/easy_localization.dart';
@@ -15,15 +15,14 @@ enum MainMenuItems{
   item_messages
 }
 
-class Controller extends GetxController{
+class MainViewController extends GetxController{
 //class Controller extends GetxService{
 //  CIMConnection connection;
   CIMDataProvider dataProvider;
   Rx<MainMenuItems> selectedItem;
 
-  AuthorisationViewModel authViewModel;
+//  AuthorisationViewModel authViewModel;
 //  ConnectionViewModel connectionViewModel;
-  PatientsScreenModel patientsScreenModel;
   RxBool authorised;
   bool isAuthorised() => authorised.value;
   void authorise(CIMUser user){
@@ -48,9 +47,8 @@ class Controller extends GetxController{
     super.onInit();
     dataProvider = CIMDataProvider();
     selectedItem = Rx(MainMenuItems.item_patients);
-    authViewModel = AuthorisationViewModel();
+//    authViewModel = AuthorisationViewModel();
 //    connection = Get.put(CIMConnection());
-    patientsScreenModel = PatientsScreenModel();
     authorised = false.obs;
   }
 }
