--- cim_protocol/lib/src/CIMRestApi.dart
@@ -1,3 +1,0 @@
-class CIMRestApi{
-  static String prepareCheckConnection() => '/checkConnection';
-}
