+++ cim_protocol/lib/src/cim_rest_api.dart
@@ -1,0 +1,6 @@
+import 'cim_user.dart';
+
+class CIMRestApi{
+  static String prepareCheckConnection() => '/checkConnection';
+  static String prepareAuthoriseUser() => '/authoriseUser';
+}
